Add --compile-css command line flag.
By default lessc simply copies @import statements referencing
CSS files verbatimly.  This default is not affected by this commit.

However if the --compile-css flag is provided on the command line,
CSS files will be processed like LESS files would.  That is to say
they are no longer externally referenced but inlined (and
optionally compressed).

Keep in mind that this leads to problems in case you've @import
statements with query strings.